HW5/requirements.txt
fastapi
uvicorn
uvicorn[standard]
pytest
pydantic
fastapi[standard]
pytest-asyncio
httpx
scikit-learn
numpy
mlflow
psycopg2-binary
alembic
asyncpg
pgmigrate
aiokafka
redis
python-dotenv
asyncpg
aiokafka